+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.5.2 (2014-03-23)
+
+  - check HTTP status code of NSURLConnection responses during xhr polling transport
+
+
+## 0.5.1 (2014-03-14)
+
+  - add method to pass cookies on handshake
+  - distinguish between handshake failed and unauthorized handshake
+
+
+## 0.5 (2014-02-21)
+
+  - finally remove deprecated delegate methods
+  - remove external JSON frameworks
+
+
 ## 0.4.1 (2013-11-18)
 
   - fix unix timecode bug in handshake url. close #118.